CFY-4698 fix connection errors in docker plugin system tests
The "permission denied" error (when connecting to a docker socket)
was caused by the fabric user not being in the docker group.
Groups are only refreshed after a new login.
Most of the time fabric would indeed use a new connection, but sometimes
it would reuse an existing one.

This makes the system test forcibly disconnect all cached fabric
connections, so when it first uses docker, it necessarily needs to
log in again.